fixed code instrumentation for convolution

+++ convolution.h
@@ -3,5 +3,5 @@
 #include <math.h>
 #include "count.h"
 
-void calc_RGB_energy(int n, int m, double* channels, double* result);
-double* padd0_image(int n, int m, double* channels);
+void calc_RGB_energy(int n, int m, int* channels, int* result);
+int* padd0_image(int n, int m, int* channels);
